move clean ct jobs
