+++ projects/_01_ingest/capacity_tracker/tests/test_clean_capacity_tracker_non_res_data.py
@@ -1,11 +1,19 @@
 import unittest
 from unittest.mock import ANY, Mock, patch
 
-import jobs.clean_capacity_tracker_non_res_data as job
-from tests.test_file_data import CapacityTrackerNonResData as Data
-from tests.test_file_schemas import CapacityTrackerNonResSchema as Schemas
+import projects._01_ingest.capacity_tracker.jobs.clean_capacity_tracker_non_res_data as job
+from projects._01_ingest.unittest_data.ingest_test_file_data import (
+    CapacityTrackerNonResData as Data,
+)
+from projects._01_ingest.unittest_data.ingest_test_file_schemas import (
+    CapacityTrackerNonResSchema as Schemas,
+)
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import PartitionKeys as Keys
+
+PATCH_PATH: str = (
+    "projects._01_ingest.capacity_tracker.jobs.clean_capacity_tracker_non_res_data"
+)
 
 
 class CapacityTrackerNonResTests(unittest.TestCase):
@@ -30,8 +38,8 @@ class MainTests(CapacityTrackerNonResTests):
             Schemas.capacity_tracker_non_res_schema,
         )
 
-    @patch("utils.utils.write_to_parquet")
-    @patch("utils.utils.read_from_parquet")
+    @patch(f"{PATCH_PATH}.utils.write_to_parquet")
+    @patch(f"{PATCH_PATH}.utils.read_from_parquet")
     def test_main(self, read_from_parquet_mock: Mock, write_to_parquet_mock: Mock):
         read_from_parquet_mock.return_value = self.test_df
