+++ projects/_01_ingest/unittest_data/ingest_test_file_data.py
@@ -618,6 +618,104 @@ class CapacityTrackerNonResData:
 
 
 @dataclass
+class CapacityTrackerCareHomeData:
+    capacity_tracker_care_home_rows = [
+        (
+            "loc 1",
+            "1",
+            "2",
+            "3",
+            "4",
+            "5",
+            "6",
+            "2024",
+            "01",
+            "01",
+            "20240101",
+            "other data",
+        ),
+    ]
+
+    remove_matching_agency_and_non_agency_rows = [
+        ("loc 1", "1", "2", "3", "4", "5", "6"),
+        ("loc 2", "1", "2", "3", "1", "5", "6"),
+        ("loc 3", "1", "2", "3", "4", "2", "6"),
+        ("loc 4", "1", "2", "3", "4", "5", "3"),
+        ("loc 5", "1", "2", "3", "1", "2", "6"),
+        ("loc 6", "1", "2", "3", "1", "5", "3"),
+        ("loc 7", "1", "2", "3", "4", "2", "3"),
+        ("loc 8", "1", "2", "3", "1", "2", "3"),
+    ]
+    expected_remove_matching_agency_and_non_agency_rows = [
+        ("loc 1", "1", "2", "3", "4", "5", "6"),
+        ("loc 2", "1", "2", "3", "1", "5", "6"),
+        ("loc 3", "1", "2", "3", "4", "2", "6"),
+        ("loc 4", "1", "2", "3", "4", "5", "3"),
+        ("loc 5", "1", "2", "3", "1", "2", "6"),
+        ("loc 6", "1", "2", "3", "1", "5", "3"),
+        ("loc 7", "1", "2", "3", "4", "2", "3"),
+    ]
+
+    create_new_columns_with_totals_rows = [
+        ("loc 1", 1, 2, 3, 40, 50, 60),
+    ]
+    expected_create_new_columns_with_totals_rows = [
+        ("loc 1", 1, 2, 3, 40, 50, 60, 6, 150, 156),
+    ]
+
+
+@dataclass
+class CapacityTrackerNonResData:
+    capacity_tracker_non_res_rows = [
+        ("loc 1", "12", "300", "2024", "01", "01", "20240101", "other data"),
+    ]
+
+
+@dataclass
+class ValidateCleanedCapacityTrackerCareHomeData:
+    ct_care_home_rows = [
+        ("1-000000001", "1", "2", "3", "4", "5", "6", "2024", "01", "01"),
+        ("1-000000002", "1", "2", "3", "4", "5", "6", "2024", "01", "01"),
+        ("1-000000001", "1", "2", "3", "4", "5", "6", "2024", "02", "01"),
+        ("1-000000002", "1", "2", "3", "4", "5", "6", "2024", "02", "01"),
+    ]
+
+    # fmt: off
+    cleaned_ct_care_home_rows = [
+        ("1-000000001", "1", "2", "3", "4", "5", "6", "2024", "01", "01", date(2024, 1, 1), 6, 15, 21),
+        ("1-000000002", "1", "2", "3", "4", "5", "6", "2024", "01", "01", date(2024, 1, 1), 6, 15, 21),
+        ("1-000000001", "1", "2", "3", "4", "5", "6", "2024", "02", "01", date(2024, 2, 1), 6, 15, 21),
+        ("1-000000002", "1", "2", "3", "4", "5", "6", "2024", "02", "01", date(2024, 2, 1), 6, 15, 21),
+    ]
+    # fmt: on
+
+    calculate_expected_size_rows = [
+        ("1-000000001", "1", "2", "3", "4", "5", "6", "2024", "01", "01"),
+    ]
+
+
+@dataclass
+class ValidateCleanedCapacityTrackerNonResData:
+    ct_non_res_rows = [
+        ("1-000000001", "1", "2", "2024", "01", "01"),
+        ("1-000000002", "1", "2", "2024", "01", "01"),
+        ("1-000000001", "1", "2", "2024", "02", "01"),
+        ("1-000000002", "1", "2", "2024", "02", "01"),
+    ]
+
+    cleaned_ct_non_res_rows = [
+        ("1-000000001", "1", "2", "2024", "01", "01", date(2024, 1, 1)),
+        ("1-000000002", "1", "2", "2024", "01", "01", date(2024, 1, 1)),
+        ("1-000000001", "1", "2", "2024", "02", "01", date(2024, 2, 1)),
+        ("1-000000002", "1", "2", "2024", "02", "01", date(2024, 2, 1)),
+    ]
+
+    calculate_expected_size_rows = [
+        ("1-000000001", "1", "2", "2024", "01", "01"),
+    ]
+
+
+@dataclass
 class IngestONSData:
     sample_rows = [
         ("Yorkshire & Humber", "Leeds", "50.10101"),
